requirements.txt
# Core dependencies for Trinity architecture
torch>=2.1.0
transformers>=4.38.0
accelerate>=0.25.0
bitsandbytes>=0.41.0
peft>=0.7.0
trl>=0.7.0

# Unsloth is installed in Colab via the notebook pip cells.
# xformers is intentionally not pinned here because its wheels are GPU/CUDA-specific
# and uv cannot resolve them for a CPU-only local environment.

# Data processing and utilities
numpy>=1.24.0
datasets>=2.14.0
scipy>=1.10.0
tqdm>=4.65.0

# Environment and API management
python-dotenv>=1.0.0

# Additional utilities for Phase 2 (Ghost Architect)
pillow>=10.0.0
sqlparse>=0.4.4
rich>=13.0.0
streamlit>=1.32.0
google-generativeai>=0.3.0
